functions distributed to different javascript files using webpack
